Add xcape command to .profile
The command remaps the left control key to trigger "escape." When
hitting control followed by any other key (e.g. ctrl-c, ctrl-s) the
effect is unchanged.

Ideally, this command would occur in conjunction with a system remaping
of the capslock key to control, so that in applications like Vim the
vast majority of operations can be completed entirely on the home row.
Otherwise, hopefully the command will be unobtrusive to the person not
using it.